+++ app/api/index.js
@@ -102,6 +102,7 @@ export type RedeemPaperVendedAdaResponse = RedeemPaperVendedAdaRequest;
 
 export type ImportKeyRequest = {
   filePath: string,
+  walletPassword: ?string,
 };
 
 export type ImportKeyResponse = Wallet;